The onus is now on the main thread to tell other threads to exit
Previously, dc_is_running would directly tell the gfx and io threads to clean up
and exit from whatever context it was called; this caused hanging issues
sometimes because gfx didn't know whether or not it was safe to call libevent
functions to wake up io_thread.

5ff310904a3ea1bb83fe82aff7c735fcdfacb93a tried to solve this with an atomic
boolean variable that tells io_thread_kick whether or not it's actually safe to
kick the io thread.  This seems to have worked because the problem immediately
went away.  However, upon closer code inspection I've realized this isn't
actually good enough because the io thread can exit between the time that
io_thread_kick checks the bool and the time that it actually calls the libevent
function to kick the io thread.

The current approach is better because only the main/emu thread will kick the
io thread during shutdown; thus we know it will only get kicked while it's still
running.